Made Window Module Platform Agnostic
- Refactored CMake scripts to better define API/Renderer subsystems;
- Added OpenGL specific Context and Window implementations;
- Temporarily added OpenGL libs to Imgui API;
- Removed API specific code from Window interface class;
