+++ cmake/vars.cmake
@@ -19,17 +19,9 @@ set(PCH_FILE ${SKELLY_PATH}/pch.cpp)
 ## Skelly libs
 set(skelly_libs 
     imgui
-
     skelly_logger
-
+    skelly_api
     skelly_render
-    # skelly_contexts
-    # skelly_api
-    # skelly_buffer
-    # skelly_shader
-    
-    
-
     skelly_window
     skelly_layer
     skelly_core
@@ -46,7 +38,9 @@ set(imgui_path
     ${VENDOR_PATH}/imgui/backends/imgui_impl_glfw.cpp
 )
 set(skelly_logger_path ${SKELLY_PATH}/utils/logger/logger.cpp)
-
+##WIP: Conditionals for environment should go here
+set(skelly_api_path ${SKELLY_PATH}/API/opengl.cpp)
+##
 set(skelly_render_path
     ${SKELLY_PATH}/utils/render/render.cpp
     ${SKELLY_PATH}/utils/render/rendererAPI.cpp
@@ -56,11 +50,6 @@ set(skelly_render_path
     ${SKELLY_PATH}/utils/render/context/renderContext.cpp
     ${SKELLY_PATH}/utils/render/shader/shader.cpp
 )
-# set(skelly_contexts_path )
-# set(skelly_buffer_path )
-# set(skelly_shader_path )
-# set(skelly_api_path )
-
 
 set(skelly_window_path ${SKELLY_PATH}/utils/window/window.cpp)
 set(skelly_layer_path ${SKELLY_PATH}/utils/layer/layer.cpp ${SKELLY_PATH}/utils/layer/layerStack.cpp)
@@ -77,24 +66,21 @@ set(imgui_includes
     ${VENDOR_PATH}/imgui/backends/
 )
 set(skelly_logger_includes ${SKELLY_PATH}/utils/logger/)
-
-set(skelly_render_includes
+set(skelly_api_includes
+    ${SKELLY_PATH}/API/
     ${SKELLY_PATH}/utils/render/
     ${SKELLY_PATH}/utils/render/context/
     ${SKELLY_PATH}/utils/render/buffer/
     ${SKELLY_PATH}/utils/render/shader/
-    ${SKELLY_PATH}/API/
+    ${SKELLY_PATH}/utils/window/
+    ${SKELLY_PATH}/utils/eventHandler/
 )
-
-# set(skelly_contexts_includes)
-# set(skelly_buffer_includes)
-# set(skelly_shader_includes)
-# set(skelly_api_includes)
-
+set(skelly_render_includes
+    ${SKELLY_PATH}/utils/render/
+)
 
 set(skelly_window_includes 
     ${SKELLY_PATH}/utils/window/
-    ${SKELLY_PATH}/utils/eventHandler/
 )
 set(skelly_layer_includes  ${SKELLY_PATH}/utils/layer/)
 
@@ -107,8 +93,7 @@ set(skelly_core_includes
 if(NOT ${PCH_ENABLED})
     set(skelly_logger_includes ${skelly_logger_includes} ${SKELLY_PATH}/ ${VENDOR_PATH}/spdlog/include/)
     set(skelly_render_includes ${skelly_render_includes} ${SKELLY_PATH}/ ${VENDOR_PATH}/imgui/)
-    # set(skelly_contexts_includes ${skelly_contexts_includes} ${SKELLY_PATH}/)
+    set(skelly_api_includes ${skelly_api_includes} ${SKELLY_PATH}/)
     set(skelly_shader_includes ${skelly_shader_includes} ${SKELLY_PATH}/)
     set(skelly_layer_includes ${skelly_layer_includes} ${SKELLY_PATH}/ ${SKELLY_PATH}/utils/eventHandler/)
-    
 endif()
