+++ skelly/CMakeLists.txt
@@ -35,23 +35,14 @@ if(${PCH_ENABLED})
     endforeach()
 endif()
 
-target_link_libraries(skelly_render PUBLIC skelly_logger glad glm)
-target_compile_definitions(skelly_render PRIVATE IMGUI_IMPL_OPENGL_LOADER_CUSTOM)
-
-# target_link_libraries(skelly_contexts PUBLIC skelly_logger glad)
-# target_link_libraries(skelly_api PUBLIC skelly_render skelly_logger skelly_buffer glad)
-# target_link_libraries(skelly_buffer PUBLIC skelly_render skelly_api)
-# target_link_libraries(skelly_shader PUBLIC skelly_logger glad)
-
-target_link_libraries(skelly_window PUBLIC skelly_render glfw)
+## WIP: conditionals should go here
+target_link_libraries(skelly_api PUBLIC skelly_logger glad glfw glm)
+target_compile_definitions(skelly_api PRIVATE IMGUI_IMPL_OPENGL_LOADER_CUSTOM)
+##
+target_link_libraries(skelly_render PUBLIC skelly_logger skelly_api)
+target_link_libraries(skelly_window PUBLIC skelly_api)
 target_compile_definitions(skelly_core PRIVATE IMGUI_IMPL_OPENGL_LOADER_CUSTOM)
 target_link_libraries(skelly_core PUBLIC imgui skelly_window skelly_render skelly_layer)
-
-#skelly_render
-#skelly_contexts
-#skelly_buffer
-#skelly_shader
-#skelly_api
 
 foreach(lib IN LISTS skelly_libs)
     target_include_directories(${lib} PUBLIC ${${lib}_includes})
