+++ skelly/utils/layer/imgui/imguiLayer.h
@@ -8,6 +8,9 @@
 #include <backends/imgui_impl_glfw.h>
 #include <backends/imgui_impl_opengl3.h>
 
+#include <glad/gl.h>
+#include <GLFW/glfw3.h>
+
 #ifndef PCH_ENABLED
     #include <layer.h>
     #include <mouseEvent.h>
